+++ features/step_definitions/cli_steps.js
@@ -1,43 +1,22 @@
 /* eslint-disable babel/new-cap */
 
 import {defineSupportCode} from '../../'
-import {execFile} from 'child_process'
 import {expect} from 'chai'
 import {normalizeText} from '../support/helpers'
-import colors from 'colors/safe'
-import fs from 'fs'
-import path from 'path'
 import stringArgv from 'string-argv'
 
-const executablePath = path.join(__dirname, '..', '..', 'bin', 'cucumber.js')
-
 defineSupportCode(function({When, Then}) {
-  When(/^I run cucumber.js(?: with `(|.+)`)?$/, {timeout: 10000}, function(args, callback) {
+  When(/^I run cucumber.js(?: with `(|.+)`)?$/, {timeout: 10000}, function(args) {
     args = stringArgv(args || '')
-    args.unshift(executablePath)
-    args.push('--backtrace', '--format', 'json:out.json')
-    const cwd = this.tmpDir
-    execFile('node', args, {cwd}, (error, stdout, stderr) => {
-      let jsonOutput = []
-      const jsonOutputPath = path.join(cwd, 'out.json')
-      if (fs.existsSync(jsonOutputPath)) {
-        const fileContent = fs.readFileSync(jsonOutputPath, 'utf8')
-        if (fileContent) {
-          jsonOutput = JSON.parse(fileContent)
-        }
-      }
-      if (this.debug) {
-        console.log(stdout + stderr) // eslint-disable-line no-console
-      }
-      this.lastRun = {
-        error,
-        jsonOutput,
-        output: colors.strip(stdout) + stderr
-      }
-      this.verifiedLastRunError = false
-      expect(this.lastRun.output).to.not.include('Unhandled rejection')
-      callback()
-    })
+    return this.run(this.localExecutablePath, args)
+  })
+
+  When(/^I run cucumber.js \(installed (locally|globally)\)$/, {timeout: 10000}, function(location) {
+    if (location === 'locally') {
+      return this.run(this.localExecutablePath, [])
+    } else {
+      return this.run(this.globalExecutablePath, [])
+    }
   })
 
   Then(/^it passes$/, function() {})
@@ -78,23 +57,6 @@ defineSupportCode(function({When, Then}) {
   Then(/^I see the help text for Cucumber$/, function() {
     const actualOutput = this.lastRun.output
     const expectedOutput = 'Usage: cucumber.js'
-    expect(actualOutput).to.include(expectedOutput)
-  })
-
-  Then(/^it suggests a "([^"]*)" step definition snippet(?: with (\d+) parameters?(?: named "([^"]*)")?)? for:$/, function (step, parameterCount, parameterName, regExp) {
-    const parameters = []
-    if (parameterName) {
-      parameters.push(parameterName)
-    }
-    else if (parameterCount) {
-      const count = parseInt(parameterCount)
-      for (let i = 1; i <= count; i += 1) {
-        parameters.push('arg' + i)
-      }
-    }
-    parameters.push('callback')
-    const expectedOutput = step + '(' + regExp + ', function (' + parameters.join(', ') + ') {\n'
-    const actualOutput = normalizeText(this.lastRun.output)
     expect(actualOutput).to.include(expectedOutput)
   })
 })