+++ src/runtime/step_runner.js
@@ -3,27 +3,36 @@ import Status from '../status'
 import StepResult from '../models/step_result'
 import Time from '../time'
 import UserCodeRunner from '../user_code_runner'
+import Promise from 'bluebird'
 
 const {beginTiming, endTiming} = Time
 
-async function run({attachmentManager, defaultTimeout, scenarioResult, step, stepDefinition, transformLookup, world}) {
+async function run({attachmentManager, defaultTimeout, scenarioResult, step, stepDefinition, parameterTypeRegistry, world}) {
   beginTiming()
-  const parameters = stepDefinition.getInvocationParameters({scenarioResult, step, transformLookup})
-  const timeoutInMilliseconds = stepDefinition.options.timeout || defaultTimeout
+  let error, result, parameters
 
-  let error, result
-  const validCodeLengths = stepDefinition.getValidCodeLengths(parameters)
-  if (_.includes(validCodeLengths, stepDefinition.code.length)) {
-    const data = await UserCodeRunner.run({
-      argsArray: parameters,
-      fn: stepDefinition.code,
-      thisArg: world,
-      timeoutInMilliseconds
-    })
-    error = data.error
-    result = data.result
-  } else {
-    error = stepDefinition.getInvalidCodeLengthMessage(parameters)
+  try {
+    parameters = await Promise.all(stepDefinition.getInvocationParameters({scenarioResult, step, parameterTypeRegistry}))
+  } catch(err) {
+    error = err
+  }
+
+  if(!error) {
+    const timeoutInMilliseconds = stepDefinition.options.timeout || defaultTimeout
+
+    const validCodeLengths = stepDefinition.getValidCodeLengths(parameters)
+    if (_.includes(validCodeLengths, stepDefinition.code.length)) {
+      const data = await UserCodeRunner.run({
+        argsArray: parameters,
+        fn: stepDefinition.code,
+        thisArg: world,
+        timeoutInMilliseconds
+      })
+      error = data.error
+      result = data.result
+    } else {
+      error = stepDefinition.getInvalidCodeLengthMessage(parameters)
+    }
   }
 
   const attachments = attachmentManager.getAll()